Add gamma

+++ bmplib.hpp
@@ -216,7 +216,7 @@ template<>
 void bitMapImage<32>::iLoveColorTable([[maybe_unused]] Color* colors, [[maybe_unused]] int size) {
 	this->colorTbl = byteArray(4, 0x00FF0000) + byteArray(4, static_cast<int>(0xFF000000)) + byteArray(4, 0x0000FF00);
 	this->bfOffBits = byteArray(4, 66);
-	this->bfSize = byteArray(4, 66 + static_cast<int>(ceil(this->w * this->h * (static_cast<double>(32) / 8))) + this->koef);
+	this->bfSize = byteArray(4, 66 + static_cast<int>(ceil((double)this->w * (double)this->h * (static_cast<double>(32) / 8))) + this->koef);
 }
 
 template<>
@@ -242,7 +242,7 @@ template<>
 void bitMapImage<24>::iLoveColorTable([[maybe_unused]] Color* colors,[[maybe_unused]] int size) {
 	this->colorTbl = byteArray();
 	this->bfOffBits = byteArray(4, 54);
-	this->bfSize = byteArray(4, 54 + static_cast<int>(ceil(this->w * this->h * (static_cast<double>(24) / 8))) + this->koef);
+	this->bfSize = byteArray(4, 54 + static_cast<int>(ceil((double)this->w * (double)this->h * (static_cast<double>(24) / 8))) + this->koef);
 }
 
 template<>
@@ -275,7 +275,7 @@ template<>
 void bitMapImage<16>::iLoveColorTable([[maybe_unused]] Color* colors, [[maybe_unused]] int size) {
 	this->colorTbl = byteArray(4, 0b0000011111100000) + byteArray(4, 0b1111100000000000) + byteArray(4, 0b0000000000011111);
 	this->bfOffBits = byteArray(4, 66);
-	this->bfSize = byteArray(4, 66 + static_cast<int>(ceil(this->w * this->h * (static_cast<double>(16) / 8))) + this->koef);
+	this->bfSize = byteArray(4, 66 + static_cast<int>(ceil((double)this->w * (double)this->h * (static_cast<double>(16) / 8))) + this->koef);
 }
 
 template<>
